Remove hooks
